# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta es la documentación de la API de Loyalty Service",
        "title": "Loyalty Service API Documentation",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/api",
    "paths": {
        "/campaigns": {
            "get": {
                "description": "Get all campaigns",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "campaigns"
                ],
                "summary": "Get all campaigns",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/campaigns/:campaignId": {
            "put": {
                "description": "Get all campaigns",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "campaigns"
                ],
                "summary": "Get all campaigns",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the campaign to update",
                        "name": "campaignId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}